+++ src/test/kotlin/uk/co/ben_gibson/git/link/url/AzureTest.kt
@@ -28,17 +28,17 @@ class AzureTest {
             Arguments.of(
                 REMOTE_BASE_URL_WITH_GIT,
                 UrlOptions.UrlOptionsFileAtBranch(FILE, BRANCH, LINE_SELECTION),
-                "https://dev.azure.com/ben-gibson/_git/test.git?version=GBmaster&path=src%2FFoo.java&line=10&lineEnd=21&lineStartColumn=1&lineEndColumn=1"
+                "https://dev.azure.com/ben-gibson/_git/test?version=GBmaster&path=src%2FFoo.java&line=10&lineEnd=21&lineStartColumn=1&lineEndColumn=1"
             ),
             Arguments.of(
                 REMOTE_BASE_URL_WITH_GIT,
                 UrlOptions.UrlOptionsFileAtBranch(FILE, BRANCH),
-                "https://dev.azure.com/ben-gibson/_git/test.git?version=GBmaster&path=src%2FFoo.java"
+                "https://dev.azure.com/ben-gibson/_git/test?version=GBmaster&path=src%2FFoo.java"
             ),
             Arguments.of(
                 REMOTE_BASE_URL_WITH_GIT,
                 UrlOptions.UrlOptionsFileAtCommit(FILE, "main", COMMIT, LINE_SELECTION),
-                "https://dev.azure.com/ben-gibson/_git/test.git?version=GCb032a0707beac9a2f24b1b7d97ee4f7156de182c&path=src%2FFoo.java&line=10&lineEnd=21&lineStartColumn=1&lineEndColumn=1"
+                "https://dev.azure.com/ben-gibson/_git/test?version=GCb032a0707beac9a2f24b1b7d97ee4f7156de182c&path=src%2FFoo.java&line=10&lineEnd=21&lineStartColumn=1&lineEndColumn=1"
             ),
             Arguments.of(
                 REMOTE_BASE_URL_WITH_GIT,
@@ -47,7 +47,7 @@ class AzureTest {
                     "main",
                     COMMIT
                 ),
-                "https://dev.azure.com/ben-gibson/_git/test.git?version=GCb032a0707beac9a2f24b1b7d97ee4f7156de182c&path=src%2Ffoo%2Fresources"
+                "https://dev.azure.com/ben-gibson/_git/test?version=GCb032a0707beac9a2f24b1b7d97ee4f7156de182c&path=src%2Ffoo%2Fresources"
             ),
             Arguments.of(
                 REMOTE_BASE_URL_WITH_GIT,
@@ -56,36 +56,36 @@ class AzureTest {
                     "main",
                     COMMIT
                 ),
-                "https://dev.azure.com/ben-gibson/_git/test.git?version=GCb032a0707beac9a2f24b1b7d97ee4f7156de182c&path=%2F"),
+                "https://dev.azure.com/ben-gibson/_git/test?version=GCb032a0707beac9a2f24b1b7d97ee4f7156de182c&path=%2F"),
             Arguments.of(
                 REMOTE_BASE_URL_WITH_GIT,
                 UrlOptions.UrlOptionsFileAtCommit(FILE, "main", COMMIT),
-                "https://dev.azure.com/ben-gibson/_git/test.git?version=GCb032a0707beac9a2f24b1b7d97ee4f7156de182c&path=src%2FFoo.java"
+                "https://dev.azure.com/ben-gibson/_git/test?version=GCb032a0707beac9a2f24b1b7d97ee4f7156de182c&path=src%2FFoo.java"
             ),
             Arguments.of(
                 REMOTE_BASE_URL_WITH_GIT,
                 UrlOptions.UrlOptionsCommit(COMMIT, "main"),
-                "https://dev.azure.com/ben-gibson/_git/test.git/commit/b032a0707beac9a2f24b1b7d97ee4f7156de182c"
+                "https://dev.azure.com/ben-gibson/_git/test/commit/b032a0707beac9a2f24b1b7d97ee4f7156de182c"
             ),
             Arguments.of(
                 REMOTE_BASE_URL_WITHOUT_GIT,
                 UrlOptions.UrlOptionsCommit(COMMIT, "main"),
-                "https://dev.azure.com/ben-gibson/_git/test.git/commit/b032a0707beac9a2f24b1b7d97ee4f7156de182c"
+                "https://dev.azure.com/ben-gibson/_git/test/commit/b032a0707beac9a2f24b1b7d97ee4f7156de182c"
             ),
             Arguments.of(
                 REMOTE_BASE_URL_WITH_COMPANY_AND_GIT,
                 UrlOptions.UrlOptionsCommit(COMMIT, "main"),
-                "https://dev.azure.com/company/project/_git/test.git/commit/b032a0707beac9a2f24b1b7d97ee4f7156de182c"
+                "https://dev.azure.com/company/project/_git/test/commit/b032a0707beac9a2f24b1b7d97ee4f7156de182c"
             ),
             Arguments.of(
                 REMOTE_BASE_URL_WITH_COMPANY_WITHOUT_GIT,
                 UrlOptions.UrlOptionsCommit(COMMIT, "main"),
-                "https://dev.azure.com/company/project/_git/test.git/commit/b032a0707beac9a2f24b1b7d97ee4f7156de182c"
+                "https://dev.azure.com/company/project/_git/test/commit/b032a0707beac9a2f24b1b7d97ee4f7156de182c"
             ),
             Arguments.of(
                 URL.fromString("https://ssh.dev.azure.com/v3/ben-gibson/test/test"),
                 UrlOptions.UrlOptionsCommit(COMMIT, "main"),
-                "https://dev.azure.com/ben-gibson/test/_git/test.git/commit/b032a0707beac9a2f24b1b7d97ee4f7156de182c"
+                "https://dev.azure.com/ben-gibson/test/_git/test/commit/b032a0707beac9a2f24b1b7d97ee4f7156de182c"
             )
         )
     }